+++ ext/duckdb/duckdb.c
@@ -42,4 +42,5 @@ Init_duckdb_native(void) {
     rbduckdb_init_duckdb_instance_cache();
     rbduckdb_init_duckdb_value_impl();
     rbduckdb_init_duckdb_scalar_function();
+    rbduckdb_init_duckdb_table_function();
 }